refactor: Abstract getGBKit helper
Allow reuse and remote unnecessary dependency injection.
